+++ packages/terminal/src/browser/terminal-widget-impl.ts
@@ -23,7 +23,7 @@ import { isOSX } from '@theia/core/lib/common';
 import { WorkspaceService } from '@theia/workspace/lib/browser';
 import { ShellTerminalServerProxy } from '../common/shell-terminal-protocol';
 import { terminalsPath } from '../common/terminal-protocol';
-import { IBaseTerminalServer } from '../common/base-terminal-protocol';
+import { IBaseTerminalServer, TerminalProcessInfo } from '../common/base-terminal-protocol';
 import { TerminalWatcher } from '../common/terminal-watcher';
 import { TerminalWidgetOptions, TerminalWidget } from './base/terminal-widget';
 import { MessageConnection } from 'vscode-jsonrpc';
@@ -34,6 +34,7 @@ import URI from '@theia/core/lib/common/uri';
 import { TerminalService } from './base/terminal-service';
 import { TerminalCopyOnSelectionHandler } from './terminal-copy-on-selection-handler';
 import { TerminalThemeService } from './terminal-theme-service';
+import { CommandLineOptions, ShellCommandBuilder } from '@theia/process/lib/common/shell-command-builder';
 
 export const TERMINAL_WIDGET_FACTORY_ID = 'terminal';
 
@@ -67,6 +68,7 @@ export class TerminalWidgetImpl extends TerminalWidget implements StatefulWidget
     @inject(TerminalService) protected readonly terminalService: TerminalService;
     @inject(TerminalCopyOnSelectionHandler) protected readonly copyOnSelectionHandler: TerminalCopyOnSelectionHandler;
     @inject(TerminalThemeService) protected readonly themeService: TerminalThemeService;
+    @inject(ShellCommandBuilder) protected readonly shellCommandBuilder: ShellCommandBuilder;
 
     protected readonly onDidOpenEmitter = new Emitter<void>();
     readonly onDidOpen: Event<void> = this.onDidOpenEmitter.event;
@@ -238,6 +240,13 @@ export class TerminalWidgetImpl extends TerminalWidget implements StatefulWidget
             return Promise.reject(new Error('terminal is not started'));
         }
         return this.shellTerminalServer.getProcessId(this.terminalId);
+    }
+
+    get processInfo(): Promise<TerminalProcessInfo> {
+        if (!IBaseTerminalServer.validateId(this.terminalId)) {
+            return Promise.reject(new Error('terminal is not started'));
+        }
+        return this.shellTerminalServer.getProcessInfo(this.terminalId);
     }
 
     get lastTouchEndEvent(): TouchEvent | undefined {
@@ -440,6 +449,10 @@ export class TerminalWidgetImpl extends TerminalWidget implements StatefulWidget
                 connection.sendRequest('write', text)
             );
         }
+    }
+
+    async executeCommand(commandOptions: CommandLineOptions): Promise<void> {
+        this.sendText(this.shellCommandBuilder.buildCommand(await this.processInfo, commandOptions) + '\n');
     }
 
     get onTerminalDidClose(): Event<TerminalWidget> {