+++ scripts/prepare-travis.js
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+
 /********************************************************************************
  * Copyright (c) 2018 TypeFox and others
  *
@@ -51,5 +52,5 @@ const startIndex = content.indexOf('# start_cache_directories') + '# start_cache
 const endIndex = content.indexOf('# end_cache_directories');
 const result = content.substr(0, startIndex) + endOfLine +
     directories.sort((d, d2) => d.localeCompare(d2)).map(d => `    - ${d}${endOfLine}`).join('') +
-    content.substr(endIndex);
+    '    ' + content.substr(endIndex);
 fs.writeFileSync(travisPath, result);